fix(core): Make project clusters caching optional
Turns out trying to cache all project clusters takes forever in larger deployments.
Rather than try to cache everything, this change makes caching opt-in project-by-project.
