+++ clouddriver-web/src/main/groovy/com/netflix/spinnaker/clouddriver/controllers/ProjectController.groovy
@@ -18,8 +18,9 @@ package com.netflix.spinnaker.clouddriver.controllers
 
 import com.netflix.spinnaker.cats.cache.Cache
 import com.netflix.spinnaker.cats.cache.CacheData
-import com.netflix.spinnaker.clouddriver.core.agent.ProjectClustersCachingAgent
+import com.netflix.spinnaker.clouddriver.core.ProjectClustersService
 import com.netflix.spinnaker.clouddriver.core.provider.agent.Namespace
+import com.netflix.spinnaker.config.ProjectClustersCachingAgentProperties
 import com.netflix.spinnaker.kork.web.exceptions.NotFoundException
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
@@ -28,29 +29,45 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
 import org.springframework.web.bind.annotation.RestController
 
+import static com.netflix.spinnaker.clouddriver.core.ProjectClustersService.ClusterModel
+
 @Slf4j
 @RestController
 @RequestMapping("/projects/{project}")
 class ProjectController {
 
   Cache cacheView
+  ProjectClustersService projectClustersService
+  ProjectClustersCachingAgentProperties projectClustersCachingAgentProperties
 
   @Autowired
-  ProjectController(Cache cacheView) {
+  ProjectController(Cache cacheView,
+                    ProjectClustersService projectClustersService,
+                    ProjectClustersCachingAgentProperties projectClustersCachingAgentProperties) {
     this.cacheView = cacheView
+    this.projectClustersService = projectClustersService
+    this.projectClustersCachingAgentProperties = projectClustersCachingAgentProperties
   }
 
   @RequestMapping(method= RequestMethod.GET, value = "/clusters")
-  List<ProjectClustersCachingAgent.ClusterModel> getClusters(@PathVariable String project) {
-    CacheData cacheData = cacheView.get(Namespace.PROJECT_CLUSTERS.ns, "v1")
-    if (cacheData == null) {
-      throw new NotFoundException("Projects not cached")
-    }
+  List<ClusterModel> getClusters(@PathVariable String project) {
+    if (projectClustersCachingAgentProperties.getAllowList().contains(project.toLowerCase())) {
+      CacheData cacheData = cacheView.get(Namespace.PROJECT_CLUSTERS.ns, "v1")
+      if (cacheData == null) {
+        throw new NotFoundException("Projects not cached")
+      }
 
-    Object clusters = cacheData.attributes.get(project)
-    if (clusters == null) {
-      throw new NotFoundException("Project not found (name: $project)")
+      Object clusters = cacheData.attributes.get(project)
+      if (clusters == null) {
+        throw new NotFoundException("Project not found (name: $project)")
+      }
+      return clusters
+    } else {
+      List<ClusterModel> clusters = projectClustersService.getProjectClusters(project.toLowerCase())
+      if (clusters == null) {
+        throw new NotFoundException("Project not found: (name: $project)")
+      }
+      return clusters
     }
-    return clusters
   }
 }
